Share Actualizado + Boton playlist, share fix + Cancion siguiente en fin
El boton de share, ahora tiene la cancion actualizada.
El boton de Agregar a Playlist, y el de Compartir, ahora estan centrados. Y son clickeables en cualquier parte.
Al terminar una cancion, pasa a la siguiente.

+++ Musica/app/src/main/java/com/example/dh/tpmusicagrupo3/View/Fragments/Detalles/GenreFragment.java
@@ -102,7 +102,7 @@ public class GenreFragment extends Fragment implements AdapterArtistaCanciones.N
                     rvCancionesPopularesGenero.setAdapter(adapterArtistaCanciones);
                     rvCancionesPopularesGenero.setLayoutManager(new LinearLayoutManager(getContext(), LinearLayoutManager.VERTICAL, false));
                 }
-            }, currentGenre.getId().intValue());
+            }, currentGenre.getId());
     }
 
     @Override
